# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestão de Avarias API",
        "description": "Documentação OpenAPI da aplicação de gestão de tickets, equipamentos, auditoria e relatórios.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registar utilizador",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "profile_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilizador criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "abc123"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar utilizador",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "abc123"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Listar notificações do utilizador",
                "operationId": "0f031f66cd8c52bf73e89fb2d050ad83",
                "responses": {
                    "200": {
                        "description": "Lista paginada de notificações"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marcar notificação como lida",
                "operationId": "58fb6750c2d309a0bc5028c22b02f9f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificação atualizada"
                    },
                    "404": {
                        "description": "Notificação não encontrada"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/notifications/test-email": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Enviar email de teste via Mailgun",
                "operationId": "c9f9ddcef1df2e8861c2e75c356531d1",
                "responses": {
                    "200": {
                        "description": "Email de teste enviado"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "X-Auth-Token": {
                "type": "apiKey",
                "name": "X-Auth-Token",
                "in": "header"
            },
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}